+++ setup.cfg
@@ -53,7 +53,7 @@ console_scripts =
 
 [options.extras_require]
 zstd = zstandard
-multivolume = multivolumefile>=0.0.5,<0.0.6
+multivolume = multivolumefile>=0.1.1,<0.2.0
 test =
       pytest
       pytest-benchmark
