+++ tox.ini
@@ -36,6 +36,7 @@ markers =
     unit: mark a test as an unit test.
     cli: mark a test as a cli test.
     benchmark: mark a test as a benchmarking.
+    misc: misc tests
 
 [testenv:mypy]
 basepython = python3.7
